Do some basic Ctrl-C handling
This changes nothing if the shell is blocking inside bolt. But
if we are currently reading a command, or in history, or otherwise
"shell located", it does.

NonInteractiveShell - CtrlC is always a terminating action

InteractiveShell - CtrlC is an interrupt action and will reset your
prompt. It will not exit the shell. (E.g., hit Ctrl-R to search history,
then hit Ctrl-C to get out of it).

This mirrors behavior in for example the iPython shell.