Fix cypher-shell.bat when run from root directory
Previously cypher-shell.bat would recursively search from the current directory
to find the cypher-shell jat files.  However if the user was, for example, at
the root of the system drive, it would search the entire drive which would take
far too long.  Instead, this commit changes the search so that it recursively
searches from the same directory of the cypher-shell.bat.  This would support
both a standalone and integrated into Neo4j situation.

Previously the cypher-shell.bat would use locate the cyper-shell jar file using
an intermediate temp file.  This commit changes the location code to use the
`FOR /F` statement to parse a fileset and create a proces level environment
variable which is later used in the java invocation.