Fallback to unformatted output in case JAnsi fails
JAnsi will fail on Alpine Linux due to its use of musl and not libc.
Thus JAnsi native components fail to load. This makes sure we fallback
to uncolored output when JAnsi fails.

In addition, when selecting appropriate runner, we fallback to the
built-in Java check which is not as good but will be correct most times.