Added version of connected Neo4j to welcome message
Before, on all versions:

```
Connected to Neo4j at bolt://localhost:7687.
Type :help for a list of available commands or :exit to exit the shell.
Note that Cypher queries must end with a semicolon.
neo4j>
```

After, on 3.1.0-M09 and above:

```
Connected to Neo4j 3.1.0-M09 at bolt://localhost:7687.
Type :help for a list of available commands or :exit to exit the shell.
Note that Cypher queries must end with a semicolon.
neo4j>
```