Fixed handling of piping on Windows
Windows was previously hard-coded to always use the interactive shell
runner. After som additional testing this turned out to be
unnecessary. Piping in Powershell works as expected and no different
in terms of detection in Java as previously thought.

This removes most of the special cases for Windows. The only remaining
part is that we avoid calling JLine's `isatty` method on Windows
because it seemed to block forever. It would always return false on
Windows anyway.

Also noticed and fixed a related issue. When input is redirected, it's
not possible to prompt for username and password when connected
because input has of course been redirected and there will be no way
for the user to actually type in a username/password. Thus when
non-interactive, connection will always fail and never prompt.