Fix loading of Ansi library on distros without libc
This fixes the following error:

```
Exception in thread "main" java.lang.NoClassDefFoundError: Could not initialize class org.fusesource.jansi.internal.CLibrary
    at org.neo4j.shell.log.AnsiLogger.isOutputInteractive(AnsiLogger.java:63)
    at org.neo4j.shell.log.AnsiLogger.<init>(AnsiLogger.java:39)
    at org.neo4j.shell.log.AnsiLogger.<init>(AnsiLogger.java:28)
    at org.neo4j.shell.Main.startShell(Main.java:56)
    at org.neo4j.shell.Main.main(Main.java:36)
```

which occurs on distros not using libc. One such example is Alpine
Linux which is the base for Neo4j docker image. Alpine uses musl.

Windows is another example, but that was already fixed by catching
`UnsatisfiedLinkError`. The surprise was that a different error was
thrown on Linux.