Don't pass along comments to Cypher
There was a bug in how we handle comments leading to bad data being sent
when piping files containing e.g.

```
//comment
:begin
return 42;
:end
```

i.e. when the a comment is preceding a command like `:begin` or `:help`.

This PR fixes that by simply ignoring comments on the client side, there
is no reason to send comments to cypher.