Stream results of autocommit queries...
When executing queries in autocommit mode, results are no longer
materialized before printing. This was causing issues for big queries
that would materialize as OOM crashes or just generally slow execution.

For --format=PLAIN this means each records is printed directly as it
arrived in the shell.

For --format=VERBOSE, we cannot simply stream, because we need to know
the widths of the columns before we can print even the column header. To
achieve this we materialize the 1000 first rows, and compute the column
widths from this sample. Any remaining rows will be streamed. If any
remaining row contains a value that overflows it's column width, that
value will either be wrapped to the next row (--wrap=true), or truncated
(--wrap=false)