Fix connection when default database is unavailable
We should be able to connect to another database or system database
even when the default database is stopped.

So we no longer use the Java driver verifyConnectivity() method
since that is hard-wired to check connectivity toward the default
database only.

Also adapt to that the DatabaseUnavailable error message could be
masked by the driver. Treat any ServiceUnavailable the same in
interactive mode.