Use db.indexes instead of showCurrentUser as ping
We still have no built-in mechanism for pinging the database, we were
using `dbms.showCurrentUser` (for system db) however this procedure is
not registered when auth is disabled so in those cases you get a

```
neo4j@neo4j> :use system
There is no procedure with the name dbms.showCurrentUser registered for this database instance. Please ensure you've spelled the procedure name correctly and that the procedure is properly deployed.
```

This PR changes to use `db.indexes` which is also what the browser use.