Support +ssc and +s protocols
This is achieved by not specifying an encryption setting in the driver,
if none has been explicitly set in Cypher Sell. The Driver would otherwise
throw an exception if +ssc or +s protocols are combined with manual set
encryption on or off.

cl:
This adds support for neo4j+s, neo4j+ssc, bolt+s, and bolt+ssc.