Check non-interactive cli arg before checking tty
Lots of checks are done if stdin is connected to a tty and somewhere in
the mess of logic a check doesn't work properly (at least on Linux) when
using --non-interactive via a shell. For instance, if the user account
requires a password being changed, it still prompts the user.

This checks the cli arg first before bothering to check the
System.console() / isatty() results. This trickles down to checks again
during the initial exception handling AuthenticationException or
Neo4jException handling.